pp2: Implement get/set mru
- update to uint16_t.
- No need to start/stop device.
- Check vs port buffer_pools, not Linux buffer_pool sizes.

Reviewed-on: http://vgitil04.il.marvell.com:8080/36019
Tested-by: iSoC Platform CI <ykjenk@marvell.com>

+++ src/drivers/ppv2/pp2_port.h
@@ -406,10 +406,10 @@ int pp2_port_set_mtu(struct pp2_port *port, uint16_t mtu);
 void pp2_port_get_mtu(struct pp2_port *port, uint16_t *mtu);
 
 /* Set MRU */
-int pp2_port_set_mru(struct pp2_port *port, uint32_t len);
+int pp2_port_set_mru(struct pp2_port *port, uint16_t len);
 
 /* Get MRU */
-void pp2_port_get_mru(struct pp2_port *port, uint32_t *len);
+void pp2_port_get_mru(struct pp2_port *port, uint16_t *len);
 
 /* Set Unicast promiscuous */
 int pp2_port_set_uc_promisc(struct pp2_port *port, uint32_t en);
